# docs/swagger.yaml
definitions:
  model.LoginCredentials:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  model.RegisterCredentials:
    properties:
      confirm_password:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  model.UserProfile:
    properties:
      avatar_path:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  utils.JSONResponse:
    properties:
      data: {}
      error:
        type: string
      status:
        type: boolean
    type: object
info:
  contact: {}
  title: Keftegram backend API
  version: "1.0"
paths:
  /api/chats/:
    get:
      consumes:
      - application/json
      description: Возвращает список чатов пользователя, ассоциированных с текущей
        сессией
      parameters:
      - default: token=xxx
        description: token
        in: header
        name: Cookie
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.JSONResponse'
      summary: Получение чатов пользователя
      tags:
      - Chat
  /api/login:
    post:
      consumes:
      - application/json
      description: Авторизовывает, аутентифицирует существующего пользователя и возвращает
        token
      parameters:
      - description: Данные для авторизации пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.LoginCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.JSONResponse'
      summary: Авторизация пользователя
      tags:
      - Auth
  /api/logout:
    delete:
      description: Завершает текущую сессию пользователя, удаляя cookie сессии
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.JSONResponse'
      summary: Выход пользователя
      tags:
      - Auth
  /api/register:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя по данным, переданным в запросе
        и возвращает token
      parameters:
      - description: Данные для регистрации пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.RegisterCredentials'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.JSONResponse'
      summary: Регистрация нового пользователя
      tags:
      - Auth
  /profile:
    get:
      description: Возвращает профиль текущего пользователя, основываясь на ID из
        контекста сессии
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.JSONResponse'
      summary: Получить профиль текущего пользователя
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Обновляет профиль текущего пользователя, включая возможность изменить
        изображение профиля
      parameters:
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: string
      - description: Данные профиля
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/model.UserProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.JSONResponse'
      summary: Обновить профиль текущего пользователя
      tags:
      - User
  /profile/{user_id}:
    get:
      description: Возвращает профиль пользователя по предоставленному ID
      parameters:
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.JSONResponse'
      summary: Получить профиль пользователя по ID
      tags:
      - User
  /uploads/{folder}/{name}:
    get:
      description: Возвращает файл из указанной папки на сервере
      parameters:
      - description: Название папки
        in: path
        name: folder
        required: true
        type: string
      - description: Имя файла
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.JSONResponse'
      summary: Получение загруженного файла
      tags:
      - Uploads
swagger: "2.0"